Redo matching
Previous parser was simplistic, and was reading Clojure terms "as is",
which caused side-effects like inability to match "0xabc" which was
failing to read as a number.

New parser/matcher are reading bracketed Clojure expressions, strings
(including unfinished, like "abc) and freeform values, thus creating
UI which feels more DWIM.

As a nice and unintended side-effect match : now highlights all keyword
keys in the state.